+++ package.json
@@ -19,7 +19,7 @@
     "react-dom": "18.2.0",
     "react-helmet-async": "1.3.0",
     "react-redux": "8.1.1",
-    "react-router-dom": "6.14.0"
+    "react-router-dom": "^6.14.0"
   },
   "devDependencies": {
     "@jedmao/redux-mock-store": "3.0.5",
