+++ src/components/film-poster.tsx
