+++ src/components/film-rating.tsx
