+++ src/components/movie-page/film-details.tsx
