+++ src/pages/movie-page/movie-page-description.tsx
@@ -1,0 +1,7 @@
+import {MoviePageBase} from './movie-page-base.tsx';
+import {FilmDescription} from '../../components/movie-page/film-description.tsx';
+import {Action} from '../../components/my-list.tsx';
+
+export const MoviePageDescription = () => (
+  <MoviePageBase content={<FilmDescription/>} activeTab='Overview' action={Action.Add}/>
+);
