+++ src/pages/movie-page/movie-page-details.tsx
@@ -1,0 +1,7 @@
+import {MoviePageBase} from './movie-page-base.tsx';
+import {FilmDetails} from '../../components/movie-page/film-details.tsx';
+import {Action} from '../../components/my-list.tsx';
+
+export const MoviePageDetails = () => (
+  <MoviePageBase content={<FilmDetails/>} activeTab='Details' action={Action.Add}/>
+);
