+++ src/pages/player.tsx
@@ -1,0 +1,5 @@
+import {Action, VideoPlayer} from '../components/video-player.tsx';
+
+export const Player = () => (
+  <VideoPlayer action={Action.Play} />
+);
